共通で使用するメソッドをapplicationControllerへ移動
