+++ app/controllers/application_controller.rb
@@ -1,4 +1,17 @@
 class ApplicationController < ActionController::Base
   protect_from_forgery with: :exception
+  helper_method :logged_in?, :current_user
+  before_action :logged_in?
+
   include SessionsHelper
+
+  private
+  
+  def current_user
+    @current_user ||= User.find_by(id: session[:user_id])
+  end
+
+  def logged_in?
+    current_user.present?
+  end
 end
