+++ app/helpers/sessions_helper.rb
@@ -3,14 +3,6 @@ module SessionsHelper
     session[:user_id] = user.id
   end
 
-  def current_user
-    @current_user ||= User.find_by(id: session[:user_id])
-  end
-
-  def logged_in?
-    current_user.present?
-  end
-
   def log_out
     session.delete(:user_id)
     @current_user = nil
